Consolidates _compute_metric overrides
The previous version of this code actually broke the mypy configuration because it
introduced overrides to _compute_metric that had invalid signatures. The goal was to
make it a bit clearer what the arguments of concreate class implementations were.
An alternative is to simply retain the **kwargs and unpack them as-needed. We can still
document these in the method docstrings, and we only lose IDE-based introspection
on types (indirectly).